fix(gateway): content-addressed paths mount on top of DNSLink
This enables us to mount /ipfs/ and /ipns/ on top of DNSLink
website in a way that prioritizes content-addressed paths.

Example:

Gateway: {
    PublicGateways: {
    ipfs.io: {
	PathPrefixes: [/ipfs, /ipns],
	UseSubdomains: false

Potential use for this is DNSLink website for IPFS Gateway that keeps
original URL (more elegant alternative to RootRedirect)
or just a regular website which uses content-addressing internally.

NoFetch could be used when loading of random CIDs is not desired.

License: MIT
Signed-off-by: Marcin Rataj <lidel@lidel.org>